wechat-mc-robot/requirements.txt
flask==2.3.3
flask-login==0.6.2
apscheduler==3.10.1
requests==2.31.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.0.1
sqlite3
python-dotenv==1.0.0